Added decorators for controlling CurrentUser access level, added multiple way to research for reservations, finished the room implementation, added jwt guard on all resources

+++ src/rooms/dto/create-room.input.ts
@@ -1,4 +1,4 @@
-import { InputType, Field, Int } from '@nestjs/graphql';
+import { InputType, Field, Float } from '@nestjs/graphql';
 import { RoomStatus, RoomType } from '@prisma/client';
 
 @InputType()
@@ -6,7 +6,7 @@ export class CreateRoomInput {
   @Field(() => RoomType)
   type: RoomType;
 
-  @Field(() => Int)
+  @Field(() => Float)
   price: number;
 
   @Field(() => RoomStatus)
